+++ src/helpers.js
@@ -15,6 +15,29 @@ if (!RegExp.escape) {
 }
 
 /**
+ * A rough equivalent of Array.prototype.flatMap, which is Node >= 11 only.
+ * This isn't a spec-compliant polyfill, just a small helper for my
+ * specific use cases.
+ */
+function flatMap(arr, mapper) {
+  if (typeof arr.flatMap === 'function') {
+    return arr.flatMap(mapper);
+  }
+  let ret = [];
+  arr.forEach((...args) => {
+    let result = mapper.call(this, ...args);
+    if (Array.isArray(result)) {
+      for (let thing of result) {
+        ret.push(thing);
+      }
+    } else {
+      ret.push(result);
+    }
+  });
+  return ret;
+}
+
+/**
  * Create regexps for matching zone names.
  * Returns an array of regexps matching the values of `matchZones`:
  * - createZoneMatchers(string) => [RegExpToMatchString]
@@ -56,7 +79,8 @@ function createZoneMatchers(matchZones) {
 function cacheKey(tzdata, config) {
   return JSON.stringify({
     version: tzdata.version,
-    zones: config.matchZones.toString(),
+    zones: String(config.matchZones),
+    countries: String(config.matchCountries),
     dates: [config.startYear, config.endYear],
   });
 }
@@ -101,6 +125,7 @@ function cacheFile(tzdata, config, cacheDirPath) {
 module.exports = {
   pluginName,
   createZoneMatchers,
+  flatMap,
   cacheDir,
   cacheFile,
 };